refactor: extract common resolve utils
