+++ crates/rolldown/src/module_loader/mod.rs
@@ -4,3 +4,4 @@ mod runtime_module_task;
 pub mod task_context;
 pub use module_loader::ModuleLoader;
 pub mod external_module_task;
+pub mod resolve_utils;
