+++ examples/basic-vue/rolldown.config.js
@@ -1,25 +1,21 @@
-import { defineConfig } from 'rolldown';
+import { defineConfig } from "rolldown";
 
 export default defineConfig({
-  input: './index.js',
-  resolve: {
-    // This needs to be explicitly set for now because oxc resolver doesn't
-    // assume default exports conditions. Rolldown will ship with a default that
-    // aligns with Vite in the future.
-    conditionNames: ['import'],
-  },
-  output: {
-    plugins: [
-      {
-        name: 'test-plugin',
-        outputOptions: function(options) {
-          options.banner = '/* banner */';
-          return options;
-        },
-      },
-    ],
-  },
-  experimental: {
-    enableComposingJsPlugins: true,
-  },
+	input: "index.js",
+	platform: "node",
+	inject: {
+		require: `data:application/javascript;pltaintext,import { createRequire } from 'module'; export const require = createRequire(import.meta.url);`,
+	},
+	plugins: [
+		{
+			name: "test",
+			resolveId(id) {
+				console.log(`id: `, id);
+			},
+			transform(code, id) {
+				console.log(`id: `, id);
+				console.log(code);
+			},
+		},
+	],
 });
